// src/enums/SyntaxValidationProvider.ts
export enum SyntaxErrorCodes {
    INVALID_TOKEN = "InvalidToken",
    INVALID_USERNAME = "InvalidUsername",
    INVALID_PASSWORD = "InvalidPassword",
    INVALID_FIELD_COUNT = "InvalidFieldCount",

    DOCUMENT_NOT_PROVIDED = "DocumentNotProvided",
    INVALID_PREFIX_PROVIDED = "InvalidPrefixProvided",
    INVALID_DOCUMENT_CONTENT_PROVIDED = "InvalidDocumentContentProvided",

    INVALID_FIRST_NAME = "InvalidFirstName",
    INVALID_SURNAME = "InvalidSurName",
    INVALID_ADDRESS = "InvalidAddress",
    INVALID_PHONE = "InvalidPhone",

    INVALID_CUT_DATE = "InvalidCutDate",
    INVALID_INSCRIPTION_DATE = "InvalidInscriptionDate",
    INVALID_MONT_AMMOUNT = "InvalidMonthAmmount",

    INVALID_AMMOUNT_PROVIDED = "InvalidAmmountProvided",
    INVALID_CURRENCY_PROVIDED = "InvalidCurrencyProvided",
    INVALID_PAYMENT_METHOD = "InvalidPaymentMethod",

    INVALID_ROLE_ACTION = "InvalidRoleAction",

    INVALID_DESCRIPTION = "InvalidDescription",

    INVALID_NAME = "InvalidName",
}